renamed some, encapsulated more behaviour in authorization.service.ts
