+++ transcendence-app/src/chat/chat.module.ts
@@ -7,9 +7,9 @@ import { ChatGateway } from './chat.gateway';
 import { ChatroomMemberService } from './chatroom/chatroom-member/chatroom-member.service';
 import { UserService } from '../user/user.service';
 import { AvatarModule } from '../shared/avatar/avatar.module';
-import { AuthorizationModule } from '../shared/casl/authorization.module';
-import { CaslAbilityFactory } from '../shared/casl/casl-ability.factory';
-import { AuthorizationService } from '../shared/casl/authorization.service';
+import { AuthorizationModule } from '../shared/authorization/authorization.module';
+import { CaslAbilityFactory } from '../shared/authorization/casl-ability.factory';
+import { AuthorizationService } from '../shared/authorization/authorization.service';
 import { IChatroomMemberRepository } from './chatroom/chatroom-member/infrastructure/db/chatroom-member.repository';
 import { ChatroomMemberPostgresRepository } from './chatroom/chatroom-member/infrastructure/db/postgres/chatroom-member.postgres.repository';
